WIP: Added executeRules in ValidationRunner
